+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Master] - 2025-03-26
 
+### Changed
+
+- MAJOR Refactor the VOF interface sharpening and reinitialization parameters and merge them in one subsection called interface regularization method. [#1467](https://github.com/chaos-polymtl/lethe/pull/1467)
+
+## [Master] - 2025-03-26
+
 ### Added
 
 - MINOR Addition of the explicit template instantiations for the PropertiesIndex of multiphysic DEM simulations. They are added to all files where there were already DEMProperties template instantiations. [#1469](https://github.com/chaos-polymtl/lethe/pull/1469)